refactor: guessCardCompanyByCardNumber 함수 사용처로 이관
